+++ ActivityScheduler.DataAccess/Managers/ActivityManager.cs
@@ -143,25 +143,6 @@ namespace ActivityScheduler.Data.Managers
             return rez;
         }
 
-        public Activity Clone(Activity sca)
-        {
-            Activity acv = new Activity();
-            acv.Id = sca.Id;
-            acv.Name = sca.Name;
-            acv.BatchId = sca.BatchId;
-            acv.ActivityId = sca.ActivityId;
-            acv.AlwaysSuccess = sca.AlwaysSuccess;
-            acv.StartTime = sca.StartTime;
-            acv.TransactionId = sca.TransactionId;
-            acv.IsDomestic = sca.IsDomestic;
-            acv.IsHub = sca.IsHub;
-            acv.ChildDelay = sca.ChildDelay;
-            acv.ParentActivities = sca.ParentActivities;
-            acv.ActivityParentRule = sca.ActivityParentRule;
-            acv.ScriptPath = sca.ScriptPath;
-            return acv;
-        }
-
         public Task<CommonOperationResult> RemoveActivity(Guid id)
         {
             var delRez = _repo.DeleteAsync(id);
@@ -196,6 +177,7 @@ namespace ActivityScheduler.Data.Managers
                         activityOriginal.StartTime == activityCompare.StartTime &&
                         activityOriginal.TransactionId == activityCompare.TransactionId &&
                         activityOriginal.IsDomestic == activityCompare.IsDomestic &&
+                        activityOriginal.IsActive == activityCompare.IsActive &&
                         activityOriginal.IsHub == activityCompare.IsHub &&
                         activityOriginal.ChildDelay == activityCompare.ChildDelay &&
                         activityOriginal.ParentActivities == activityCompare.ParentActivities &&
@@ -204,6 +186,25 @@ namespace ActivityScheduler.Data.Managers
 
             return rez;
 
+        }
+        public Activity Clone(Activity source)
+        {
+            Activity acv = new Activity();
+            acv.Id = source.Id;
+            acv.BatchId = source.BatchId;
+            acv.Name = source.Name;
+            acv.ActivityId = source.ActivityId;
+            acv.AlwaysSuccess = source.AlwaysSuccess;
+            acv.StartTime = source.StartTime;
+            acv.IsActive = source.IsActive;
+            acv.TransactionId = source.TransactionId;
+            acv.IsDomestic= source.IsDomestic;
+            acv.IsHub= source.IsHub;
+            acv.ChildDelay = source.ChildDelay;
+            acv.ParentActivities = source.ParentActivities;
+            acv.ActivityParentRule= source.ActivityParentRule;
+            acv.ScriptPath = source.ScriptPath;
+            return acv;
         }
     }
 }